Разделяет модуль data на модули
